+++ src/textual/app.py
@@ -868,7 +868,7 @@ class App(Generic[ReturnType], DOMNode):
         """
 
         bindings_map: dict[str, tuple[DOMNode, Binding]] = {}
-        for namespace, bindings in self._binding_chain:
+        for namespace, bindings in self._modal_binding_chain:
             for key, binding in bindings.keys.items():
                 if existing_key_and_binding := bindings_map.get(key):
                     _, existing_binding = existing_key_and_binding