+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.61.0] - Unreleased
+
+### Added
+
+- Added App.get_default_screen
+
 ## [0.60.1] - 2024-05-15
 
 ### Fixed